ldn: Rename RyuLdn to LdnRyu
